Overhaul for Character, Camera and Input.
Made new, functioning scripts for the Character and Camera, and a demo scene for split-screen multiplayer. Completely redid the input and made separate axis for Player 1 and Player 2, in a way that the PlayerNumber can easily be used to determine which axis to use in code (eg. for horizontal movement of a specific player, you would use Input.GetAxis($"Movement Horizontal {PlayerNumber}") )